requirements.txt
playwright~=1.54.0
selectolax~=0.3.33
python-dotenv~=1.1.1